make the binary dictionary builder work for synth files

+++ languagetool-standalone/src/test/java/org/languagetool/dev/SynthDictionaryBuilderTest.java
@@ -24,31 +24,40 @@ import org.languagetool.Language;
 
 import java.io.File;
 
-public class POSDictionaryBuilderTest extends DictionaryBuilderTestHelper {
+public class SynthDictionaryBuilderTest extends DictionaryBuilderTestHelper {
 
   @Test
   @Ignore("for interactive use only")
-  public void testExportAndImport() throws Exception {
+  public void testExportPosDictAndCreateSynth() throws Exception {
     for (Language language : Language.REAL_LANGUAGES) {
+      if (language.getShortName().equals("pl")) {
+        System.out.println("WARN: skipping Polish. TODO: make it work for Polish, too (problem: synth file is too large: 36MB!)");  // TODO
+        continue;
+      }
       String langCode = language.getShortName();
       File dir = new File("./languagetool-language-modules/" + langCode + "/src/main/resources/org/languagetool/resource/" + langCode);
+      File oldBinarySynthFile = new File(dir, language.getName().toLowerCase() + "_synth.dict");
+      if (!oldBinarySynthFile.exists()) {
+        System.out.println("Ignoring " + language + ", no synth file found");
+        continue;
+      }
       File oldBinaryFile = new File(dir, language.getName().toLowerCase() + ".dict");
-      File infoFile = new File(dir, language.getName().toLowerCase() + ".info");
+      File infoFile = new File(dir, language.getName().toLowerCase() + "_synth.info");
       File exportFile = exportDictionaryContents(oldBinaryFile);
       if (exportFile.length() == 0) {
         System.out.println("Zero-size output for " + language + ", skipping dictionary generation");
         exportFile.delete();
         continue;
       }
-      POSDictionaryBuilder builder = new POSDictionaryBuilder(infoFile);
-      File newBinaryFile = builder.build(exportFile);
+      SynthDictionaryBuilder builder = new SynthDictionaryBuilder(infoFile);
+      File newBinarySynthFile = builder.build(exportFile, infoFile);
       exportFile.delete();
-      System.out.println(language + " old binary file size: " + oldBinaryFile.length() + " bytes (" + oldBinaryFile.getName() + ")");
-      System.out.println(language + " new binary file size: " + newBinaryFile.length() + " bytes (" + newBinaryFile.getAbsolutePath() + ")");
+      System.out.println(language + " old binary file size: " + oldBinarySynthFile.length() + " bytes (" + oldBinarySynthFile.getName() + ")");
+      System.out.println(language + " new binary file size: " + newBinarySynthFile.length() + " bytes (" + newBinarySynthFile.getAbsolutePath() + ")");
       // comment in to copy the new files over the old ones:
-      /*boolean b = newBinaryFile.renameTo(oldBinaryFile);
+      /*boolean b = newBinarySynthFile.renameTo(oldBinarySynthFile);
       if (!b) {
-        throw new RuntimeException("Could not rename" + newBinaryFile.getAbsolutePath() + " to " + oldBinaryFile.getCanonicalPath());
+        throw new RuntimeException("Could not rename " + newBinarySynthFile.getAbsolutePath() + " to " + oldBinarySynthFile.getCanonicalPath());
       }*/
       System.out.println("");
     }
